requirements.txt
Pillow==9.5.0
Pillow==10.0.0
PyQt5==5.15.9
PyQt5_sip==12.12.1
pyserial==3.5
